+++ libraries/runningAngle/test/unit_test_001.cpp
@@ -45,27 +45,35 @@ unittest_teardown()
 }
 
 
+unittest(test_constants)
+{
+  assertEqualFloat(0.800, RA_DEFAULT_WEIGHT, 0.0001);
+  assertEqualFloat(0.001, RA_MIN_WEIGHT,     0.0001);
+  assertEqualFloat(1.000, RA_MAX_WEIGHT,     0.0001);
+}
+
+
 unittest(test_constructor_1)
 {
   runningAngle heading(runningAngle::DEGREES);
-  assertEqualFloat(0.80, heading.getWeight(), 0.0001);
-  assertEqualFloat(0, heading.getAverage(), 0.0001);
+  assertEqualFloat(0.80, heading.getWeight(),  0.0001);
+  assertEqualFloat(0.00, heading.getAverage(), 0.0001);
 }
 
 
 unittest(test_constructor_2)
 {
   runningAngle heading(runningAngle::RADIANS);
-  assertEqualFloat(0.80, heading.getWeight(), 0.0001);
-  assertEqualFloat(0, heading.getAverage(), 0.0001);
+  assertEqualFloat(0.80, heading.getWeight(),  0.0001);
+  assertEqualFloat(0.00, heading.getAverage(), 0.0001);
 }
 
 
 unittest(test_constructor_3)
 {
   runningAngle heading(runningAngle::GRADIANS);
-  assertEqualFloat(0.80, heading.getWeight(), 0.0001);
-  assertEqualFloat(0, heading.getAverage(), 0.0001);
+  assertEqualFloat(0.80, heading.getWeight(),  0.0001);
+  assertEqualFloat(0.00, heading.getAverage(), 0.0001);
 }
 
 
@@ -90,14 +98,17 @@ unittest(test_weight)
   runningAngle heading(runningAngle::DEGREES);
   assertEqualFloat(0.80, heading.getWeight(), 0.0001);
 
-  heading.setWeight(0.85);
+  assertTrue(heading.setWeight(0.85));
   assertEqualFloat(0.85, heading.getWeight(), 0.0001);
 
-  heading.setWeight(2);
+  assertFalse(heading.setWeight(2));
   assertEqualFloat(1, heading.getWeight(), 0.0001);
 
-  heading.setWeight(-5);
+  assertFalse(heading.setWeight(-5));
   assertEqualFloat(0.001, heading.getWeight(), 0.0001);
+
+  assertTrue(heading.setWeight());    // use default
+  assertEqualFloat(0.80, heading.getWeight(), 0.0001);
 
   fprintf(stderr, "\treset()\n");
   heading.reset();
@@ -108,6 +119,8 @@ unittest(test_weight)
 unittest(test_wrap)
 {
   runningAngle heading(runningAngle::DEGREES);
+
+  heading.setMode0();
 
   assertEqualFloat(0, heading.wrap(0), 0.0001);
   assertEqualFloat(0, heading.wrap(360), 0.0001);
@@ -125,6 +138,35 @@ unittest(test_wrap)
 }
 
 
+unittest(test_mode_0)
+{
+  runningAngle heading(runningAngle::DEGREES);
+
+  for (int i = 0; i < 360; i+= 20)
+  {
+    heading.reset();
+    heading.setMode0();
+    if (i < 180) assertEqualFloat(i, heading.add(i), 0.0001);
+    else         assertEqualFloat(i-360, heading.add(i), 0.0001);
+  }
+}
+
+
+unittest(test_mode_1)
+{
+  runningAngle heading(runningAngle::DEGREES);
+
+  for (int i = 0; i < 360; i+= 20)
+  {
+    heading.reset();
+    heading.setMode1();
+    assertEqualFloat(i, heading.add(i), 0.0001);
+  }
+}
+
+
 unittest_main()
 
-// --------
+
+//  -- END OF FILE --
+